fix(ci): console checks — oversized module and stale fixture broke gates
